requirements.txt
fastapi
uvicorn
python-multipart
sqlmodel
pdfplumber
google-generativeai
pandas
openpyxl
